resume/requirements.txt
fastapi
uvicorn
pdfplumber
python-multipart
requests
python-dotenv